+++ app/src/main/java/com/konkuk/medicarecall/ui/login_info/screen/LoginStartScreen.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.konkuk.medicarecall.R
 import com.konkuk.medicarecall.ui.login_info.uistate.LoginUiState
-import com.konkuk.medicarecall.ui.login_info.component.CTAButton
+import com.konkuk.medicarecall.ui.component.CTAButton
 import com.konkuk.medicarecall.ui.login_info.viewmodel.LoginViewModel
 import com.konkuk.medicarecall.ui.model.CTAButtonType
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
